pyproject.toml
[tool.poetry]
name = "marine-mp"
version = "0.1.0"
description = ""
authors = ["Alan Huynh <hdmquan@outlook.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
numpy = "^2.2.3"
h5py = "^3.13.0"
pyhdf = "^0.11.6"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
tqdm = "^4.67.1"
holoviews = "^1.20.1"
geoviews = "^1.14.0"
panel = "^1.6.1"
bokeh = "^3.6.3"
jupyter-bokeh = "^4.0.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
